+++ src/app/app.component.ts
@@ -1,9 +1,16 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss'],
 })
-export class AppComponent {
+export class AppComponent implements AfterViewInit {
+    @ViewChild('appTitle') appTitle: ElementRef<HTMLInputElement>;
+
+    constructor() {}
+
+    ngAfterViewInit() {
+        this.appTitle.nativeElement.textContent = 'Shop';
+    }
 }