+++ src/app/cart/cart-list/cart-list.component.spec.ts
@@ -1,21 +1,21 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { FirstComponent } from './first.component';
+import { CartListComponent } from './cart-list.component';
 import { SharedModule } from 'src/app/shared/shared.module';
 
-describe('FirstComponent', () => {
-    let component: FirstComponent;
-    let fixture: ComponentFixture<FirstComponent>;
+describe('CartListComponent', () => {
+    let component: CartListComponent;
+    let fixture: ComponentFixture<CartListComponent>;
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({
-            declarations: [FirstComponent],
+            declarations: [CartListComponent],
             imports: [SharedModule],
         }).compileComponents();
     }));
 
     beforeEach(() => {
-        fixture = TestBed.createComponent(FirstComponent);
+        fixture = TestBed.createComponent(CartListComponent);
         component = fixture.componentInstance;
         fixture.detectChanges();
     });