+++ src/app/core/models/index.ts
@@ -1,0 +1,5 @@
+import { ProductModel } from './product.model';
+import { ProductOrderModel } from './product-order.model';
+import { Category } from './product-category.model';
+
+export { ProductModel, ProductOrderModel, Category };