+++ src/app/core/models/product.model.ts
@@ -1,10 +1,16 @@
 import { Category } from './product-category.model';
 
-export interface ProductModel {
-    id: number;
-    name: string;
-    description: string;
-    price: number;
-    isAvailable: boolean;
-    category: Category;
+export class ProductModel {
+    constructor(
+        public id: number,
+        public name: string,
+        public description: string,
+        public price: number,
+        public category: Category,
+        public availableCount: number
+    ) {}
+
+    public isAvailable(): boolean {
+        return this.availableCount > 0;
+    }
 }