+++ src/app/core/services/index.ts
@@ -1,0 +1,4 @@
+import { CartService } from './cart/cart.service';
+import { ProductCommunicatorService } from './product-communicator/product-communicator.service';
+
+export { CartService, ProductCommunicatorService };