+++ src/app/header/header.component.spec.ts
@@ -1,19 +1,21 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CartComponent } from './cart.component';
+import { HeaderComponent } from './header.component';
+import { SharedModule } from 'src/app/shared/shared.module';
 
-describe('CartComponent', () => {
-    let component: CartComponent;
-    let fixture: ComponentFixture<CartComponent>;
+describe('HeaderComponent', () => {
+    let component: HeaderComponent;
+    let fixture: ComponentFixture<HeaderComponent>;
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({
-            declarations: [CartComponent],
+            declarations: [HeaderComponent],
+            imports: [SharedModule],
         }).compileComponents();
     }));
 
     beforeEach(() => {
-        fixture = TestBed.createComponent(CartComponent);
+        fixture = TestBed.createComponent(HeaderComponent);
         component = fixture.componentInstance;
         fixture.detectChanges();
     });