+++ src/app/product/product.component.spec.ts
@@ -1,6 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProductComponent } from './product.component';
+import { ProductModel, Category } from 'src/app/core/models';
+import { SharedModule } from 'src/app/shared/shared.module';
 
 describe('ProductComponent', () => {
     let component: ProductComponent;
@@ -9,6 +11,7 @@ describe('ProductComponent', () => {
     beforeEach(async(() => {
         TestBed.configureTestingModule({
             declarations: [ProductComponent],
+            imports: [SharedModule],
         }).compileComponents();
     }));
 
@@ -16,7 +19,7 @@ describe('ProductComponent', () => {
         fixture = TestBed.createComponent(ProductComponent);
         component = fixture.componentInstance;
 
-        component.product = {} as any;
+        component.product = new ProductModel(1, 'a', 'b', 10, Category.C1, 5);
         fixture.detectChanges();
     });